+++ data/models/ExpDistGenerator.java
@@ -5,12 +5,12 @@ import java.util.Random;
 public class ExpDistGenerator {
     private Random _randomArrival;
     private Random _randomService;
-    private float _lambdaArrival;
-    private float _lambdaService;
+    private double _lambdaArrival;
+    private double _lambdaService;
     private long _seedArrival;
     private long _seedService;
 
-    public ExpDistGenerator(float lambdaArrival, float lambdaService, long seedArrival, long seedService) {
+    public ExpDistGenerator(double lambdaArrival, double lambdaService, long seedArrival, long seedService) {
         _lambdaArrival = lambdaArrival;
         _lambdaService = lambdaService;
         _seedArrival = seedArrival;
@@ -19,12 +19,12 @@ public class ExpDistGenerator {
         _randomService = new Random(_seedService);
     }
 
-    public float nextArrivalTime() {
-        return (float) (-Math.log(1 - _randomArrival.nextFloat()) / _lambdaArrival);
+    public double nextArrivalTime() {
+        return (double) (-Math.log(1 - _randomArrival.nextFloat()) / _lambdaArrival);
     }
 
-    public float nextServiceTime() {
-        return (float) (-Math.log(1 - _randomService.nextFloat()) / _lambdaService);
+    public double nextServiceTime() {
+        return (-Math.log(1 - _randomService.nextFloat()) / _lambdaService);
     }
 
     @Override
