+++ simulation/models/EndOfExecutionEventValue.java
@@ -3,7 +3,7 @@ package simulation.models;
 public class EndOfExecutionEventValue extends EventValue {
     private int _serverNumber;
 
-    public EndOfExecutionEventValue(int categoryNumber, int serverNumber, float serviceTime) {
+    public EndOfExecutionEventValue(int categoryNumber, int serverNumber, double serviceTime) {
         _categoryNumber = categoryNumber;
         _serverNumber = serverNumber;
         _serviceTime = serviceTime;
